admin/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
mysqlclient==2.2.1
django-mysql==4.12.0
django-cors-headers==4.3.1
pika==1.3.2